requirements.txt
sqlalchemy==2.0.29
requests==2.31.0
python-dotenv==1.0.1
loguru==0.7.2
pydantic==2.6.4
tenacity==8.2.3
python-docx==1.1.0
openpyxl==3.1.2
pandas==2.2.2
PyPDF2==3.0.1
pillow==10.3.0
uvicorn==0.27.1
fastapi==0.110.0
pytest==7.4.4
Flask==2.3.3
flask-cors==4.0.0
